IOS-7105 Fix review 1
IOS-7105 Update SendFeeInteractor.swift

IOS-7105 Update SendFeeInteractor.swift

IOS-7105 Fix review 2

IOS-7105 Fix review 3

IOS-7105 Fix review 3
