+++ Tangem/Resources/Localizations/de.lproj/Localizable.strings
@@ -614,28 +614,16 @@
 "staking_details_rewards_to_claim" = "Rewards to claim: %s";
 "staking_details_title" = "Staking %@";
 "staking_details_unbonding_period" = "Entbindungsdauer";
-"staking_details_unbonding_period" = "Entbindungsdauer";
-"staking_details_unbonding_period_info" = "The period you must wait after requesting to withdraw funds from staking before the tokens become available.";
 "staking_details_unbonding_period_info" = "The period you must wait after requesting to withdraw funds from staking before the tokens become available.";
 "staking_details_warmup_period" = "Aufwärmphase";
-"staking_details_warmup_period" = "Aufwärmphase";
-"staking_details_warmup_period_info" = "The allocated time for activating participation in staking.";
 "staking_details_warmup_period_info" = "The allocated time for activating participation in staking.";
 "staking_native" = "Native staking";
-"staking_native" = "Native staking";
-"staking_notification_earn_rewards_text" = "Staking allow you to earn %1@. Your staking rewards arrive every ~%2@ days.";
 "staking_notification_earn_rewards_text" = "Staking allow you to earn %1@. Your staking rewards arrive every ~%2@ days.";
 "staking_notification_earn_rewards_title" = "Earn staking rewards";
-"staking_notification_earn_rewards_title" = "Earn staking rewards";
-"staking_rewards" = "Belohnungen";
 "staking_rewards" = "Belohnungen";
 "staking_stake_more" = "Stake more";
-"staking_stake_more" = "Stake more";
-"staking_unstaked" = "Unstacked";
 "staking_unstaked" = "Unstacked";
 "staking_unstaked_footer" = "Check unstaked to claim your assets";
-"staking_unstaked_footer" = "Check unstaked to claim your assets";
-"staking_validator" = "Validator";
 "staking_validator" = "Validator";
 "story_awe_description" = "Store your crypto assets secure while keeping private keys contained in your card";
 "story_awe_title" = "Revolutionary Hardware Wallet";
