+++ Tangem/Resources/Localizations/fr.lproj/Localizable.strings
@@ -614,28 +614,16 @@
 "staking_details_rewards_to_claim" = "Récompenses à réclamer : %s";
 "staking_details_title" = "Staking %@";
 "staking_details_unbonding_period" = "Période de détachement";
-"staking_details_unbonding_period" = "Période de détachement";
-"staking_details_unbonding_period_info" = "La période que vous devez attendre après avoir demandé le retrait des fonds du staking avant que les jetons ne soient disponibles.";
 "staking_details_unbonding_period_info" = "La période que vous devez attendre après avoir demandé le retrait des fonds du staking avant que les jetons ne soient disponibles.";
 "staking_details_warmup_period" = "Période d'échauffement";
-"staking_details_warmup_period" = "Période d'échauffement";
-"staking_details_warmup_period_info" = "Le temps imparti pour activer la participation au staking.";
 "staking_details_warmup_period_info" = "Le temps imparti pour activer la participation au staking.";
 "staking_native" = "Native staking";
-"staking_native" = "Native staking";
-"staking_notification_earn_rewards_text" = "Le staking vous permet d'en gagner %1@. Vos récompenses de staking arrivent tous les ~%2@ jours.";
 "staking_notification_earn_rewards_text" = "Le staking vous permet d'en gagner %1@. Vos récompenses de staking arrivent tous les ~%2@ jours.";
 "staking_notification_earn_rewards_title" = "Gagnez des récompenses de staking";
-"staking_notification_earn_rewards_title" = "Gagnez des récompenses de staking";
-"staking_rewards" = "Récompenses";
 "staking_rewards" = "Récompenses";
 "staking_stake_more" = "Staker plus";
-"staking_stake_more" = "Staker plus";
-"staking_unstaked" = "Non-staké";
 "staking_unstaked" = "Non-staké";
 "staking_unstaked_footer" = "Vérifiez non-stakés pour réclamer vos actifs";
-"staking_unstaked_footer" = "Vérifiez non-stakés pour réclamer vos actifs";
-"staking_validator" = "Validateur";
 "staking_validator" = "Validateur";
 "story_awe_description" = "Store your crypto assets secure while keeping private keys contained in your card";
 "story_awe_title" = "Revolutionary Hardware Wallet";
